Load rules from file at runtime
The file format is based on a small subset of the YAML language. Each
rule is a yaml list member with any of the following values:

class:    string
instance: string
title:    string
tag:      int
monitor:  int
floating: bool
terminal: bool
swallow:  bool

Everything is 0 by default except for monitor which is -1. The tag
value is automatically shifted N - 1 bits to the left.

Example for the default rules in dwm:
- class: "Gimp"
  floating: true

- class: "Firefox"
  tag: 9