Allow matching rules with regexes, implement proper quote escaping
Rules will match by regex instead of string comparison if the string is
unquoted and delimited by '/'. All quotes can now be escaped by doubling them
and backslashes cannot be used to escape single quotes. '#' needs to be
escaped with a backslash if inside a regex and preceded by a ' ' to avoid
being commented.